# ENH: optimize: release the GIL while computing the LSAP solution