#### What does this implement/fix?

The LSAP (linear sum assignment problem) routine does the bulk of its work in solve_rectangular_linear_sum_assignment() using C-style pointers for data access after unwrapping the numpy arrays, so we can release the GIL during this computation. This permits threading the computation of independent LSAP problems.

The first commit is the change to the C-code to drop the GIL, and the second commit adds a timing benchmark test for computing the solution to independent LSAPs. I don't mind dropping the second commit if it is deemed to be unnecessary.

The timing results on my machine without the first commit are:

```
$ python dev.py bench -t optimize_lap.ParallelLinearAssignment

[  0.00%] ·· Benchmarking existing-py_node_bud213_shared_envs_scipy_20230706_bin_python
[ 50.00%] ··· Running (optimize_lap.ParallelLinearAssignment.time_evaluation--).
[100.00%] ··· optimize_lap.ParallelLinearAssignment.time_evaluation                                                                                                                                                                                                                                                        ok
[100.00%] ··· ========= =============
               threads               
              --------- -------------
                  1      8.22±0.04ms 
                  2       8.44±0.1ms 
                  4      8.68±0.09ms 
              ========= =============
```

And with the first commit, this becomes:

```
$ python dev.py bench -t optimize_lap.ParallelLinearAssignment

[  0.00%] ·· Benchmarking existing-py_node_bud213_shared_envs_scipy_20230706_bin_python
[ 50.00%] ··· Running (optimize_lap.ParallelLinearAssignment.time_evaluation--).
[100.00%] ··· optimize_lap.ParallelLinearAssignment.time_evaluation                                                                                                                                                                                                                                                        ok
[100.00%] ··· ========= =============
               threads               
              --------- -------------
                  1      7.77±0.07ms 
                  2       4.14±0.1ms 
                  4       2.86±0.4ms 
              ========= =============
```