@rgommers 2023-07-12
> I don't mind dropping the second commit if it is deemed to be unnecessary.

Thanks for asking, I think it's good to keep - it's pretty fast to run, and makes it easy to check the performance gain if needed in the future.